break out IdentitySerializer
