+++ api/serializers/__init__.py
@@ -43,3 +43,4 @@ from .allocation_serializer import AllocationSerializer
 from .allocation_request_serializer import AllocationRequestSerializer
 from .quota_serializer import QuotaSerializer
 from .quota_request_serializer import QuotaRequestSerializer
+from .identity_serializer import IdentitySerializer
