+++ api/serializers/identity_serializer.py
@@ -1,0 +1,19 @@
+from core.models.identity import Identity
+from rest_framework import serializers
+from .quota_serializer import QuotaSerializer
+from .allocation_serializer import AllocationSerializer
+
+
+class IdentitySerializer(serializers.ModelSerializer):
+    created_by = serializers.CharField(source='creator_name')
+    credentials = serializers.Field(source='get_credentials')
+    id = serializers.Field(source='uuid')
+    provider_id = serializers.Field(source='provider_uuid')
+    quota = QuotaSerializer(source='get_quota')
+    allocation = AllocationSerializer(source='get_allocation')
+    membership = serializers.Field(source='get_membership')
+
+    class Meta:
+        model = Identity
+        fields = ('id', 'created_by', 'provider_id', 'credentials',
+                  'membership', 'quota', 'allocation')
